move the query snapshotting logic into fake_query_with_parent
