+++ lib/src/fake_converted_query.dart
@@ -2,15 +2,14 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 
 import 'converter.dart';
-import 'fake_query_interface.dart';
-import 'mock_query.dart';
+import 'fake_query_with_parent.dart';
 import 'mock_query_snapshot.dart';
 
 // ignore: subtype_of_sealed_class
 /// A converted query. It should always be the last query in the chain, so we
 /// don't need to implement where, startAt, ..., withConverter.
-class FakeConvertedQuery<T extends Object?> implements QueryWithParent<T> {
-  final QueryWithParent _nonConvertedParentQuery;
+class FakeConvertedQuery<T extends Object?> extends FakeQueryWithParent<T> {
+  final FakeQueryWithParent _nonConvertedParentQuery;
   final Converter<T> _converter;
 
   FakeConvertedQuery(this._nonConvertedParentQuery, this._converter)
@@ -31,16 +30,7 @@ class FakeConvertedQuery<T extends Object?> implements QueryWithParent<T> {
   }
 
   @override
-  Stream<QuerySnapshot<T>> snapshots({bool includeMetadataChanges = false}) {
-    QuerySnapshotStreamManager().register<T>(this);
-    final controller =
-        QuerySnapshotStreamManager().getStreamController<T>(this);
-    controller.addStream(Stream.fromFuture(get()));
-    return controller.stream.distinct(snapshotEquals);
-  }
-
-  @override
-  QueryWithParent? get parentQuery => _nonConvertedParentQuery;
+  FakeQueryWithParent? get parentQuery => _nonConvertedParentQuery;
 
   @override
   dynamic noSuchMethod(Invocation invocation) => super.noSuchMethod(invocation);
