+++ lib/src/mock_query.dart
@@ -3,13 +3,12 @@ import 'dart:math';
 
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:cloud_firestore_platform_interface/cloud_firestore_platform_interface.dart';
-import 'package:collection/collection.dart';
 import 'package:flutter/services.dart';
 import 'package:quiver/core.dart';
 
 import 'converter.dart';
 import 'fake_converted_query.dart';
-import 'fake_query_interface.dart';
+import 'fake_query_with_parent.dart';
 import 'mock_query_platform.dart';
 import 'mock_query_snapshot.dart';
 
@@ -17,7 +16,7 @@ typedef _QueryOperation<T extends Object?> = List<DocumentSnapshot<T>> Function(
     List<DocumentSnapshot<T>> input);
 
 // ignore: subtype_of_sealed_class
-class MockQuery<T extends Object?> implements QueryWithParent<T> {
+class MockQuery<T extends Object?> extends FakeQueryWithParent<T> {
   /// Previous query in a Firestore query chain. Null if this instance is a
   /// collection reference. A query chain always starts with a collection
   /// reference, which does not have a previous query.
@@ -47,15 +46,6 @@ class MockQuery<T extends Object?> implements QueryWithParent<T> {
     final parentQueryResult = await _parentQuery!.get(options);
     final docs = _operation!(parentQueryResult.docs);
     return MockQuerySnapshot<T>(docs);
-  }
-
-  @override
-  Stream<QuerySnapshot<T>> snapshots({bool includeMetadataChanges = false}) {
-    QuerySnapshotStreamManager().register<T>(this);
-    final controller =
-        QuerySnapshotStreamManager().getStreamController<T>(this);
-    controller.addStream(Stream.fromFuture(get()));
-    return controller.stream.distinct(snapshotEquals);
   }
 
   @override
@@ -370,105 +360,5 @@ class MockQuery<T extends Object?> implements QueryWithParent<T> {
   }
 
   @override
-  QueryWithParent? get parentQuery => _parentQuery;
-}
-
-class QuerySnapshotStreamManager {
-  static QuerySnapshotStreamManager? _instance;
-
-  factory QuerySnapshotStreamManager() =>
-      _instance ??= QuerySnapshotStreamManager._internal();
-
-  QuerySnapshotStreamManager._internal();
-  final Map<String, Map<QueryWithParent, StreamController<QuerySnapshot>>>
-      _streamCache = {};
-
-  void clear() {
-    for (final queryToStreamController in _streamCache.values) {
-      for (final streamController in queryToStreamController.values) {
-        streamController.close();
-      }
-    }
-    _streamCache.clear();
-  }
-
-  /// Recursively finds the base collection path.
-  String _getBaseCollectionPath(QueryWithParent query) {
-    // In theory retrieveParentPath should stop at the collection reference.
-    // So _parentQuery can never be null.
-    assert(query.parentQuery != null);
-    if (query.parentQuery is CollectionReference) {
-      return (query.parentQuery as CollectionReference).path;
-    } else {
-      return _getBaseCollectionPath(query.parentQuery!);
-    }
-  }
-
-  void register<T>(QueryWithParent query) {
-    final path = _getBaseCollectionPath(query);
-    if (!_streamCache.containsKey(path)) {
-      _streamCache[path] = {};
-    }
-    _streamCache[path]!.putIfAbsent(
-        query, () => StreamController<QuerySnapshot<T>>.broadcast());
-  }
-
-  void unregister(QueryWithParent query) {
-    final path = _getBaseCollectionPath(query);
-    final pathCache = _streamCache[path];
-    if (pathCache == null) {
-      return;
-    }
-    final controller = pathCache.remove(query);
-    controller!.close();
-  }
-
-  StreamController<QuerySnapshot<T>> getStreamController<T>(
-      QueryWithParent query) {
-    final path = _getBaseCollectionPath(query);
-    final pathCache = _streamCache[path];
-    // Before calling `getStreamController(query)`, one should have called
-    // `register(query)` beforehand, so pathCache should never be null.
-    assert(pathCache != null);
-    final streamController = pathCache![query]!;
-    if (streamController is! StreamController<QuerySnapshot<T>>) {
-      throw UnimplementedError();
-    }
-    return streamController;
-  }
-
-  void fireSnapshotUpdate(String path) {
-    final exactPathCache = _streamCache[path];
-    if (exactPathCache != null) {
-      for (final query in exactPathCache.keys) {
-        if (exactPathCache[query]!.hasListener) {
-          query.get().then(exactPathCache[query]!.add);
-        }
-      }
-    }
-
-    if (path.contains('/')) {
-      fireSnapshotUpdate(path.split('/').first);
-    }
-  }
-}
-
-final _unorderedDeepEquality = const DeepCollectionEquality.unordered();
-
-bool snapshotEquals(snapshot1, snapshot2) {
-  if (snapshot1.docs.length != snapshot2.docs.length) {
-    return false;
-  }
-
-  for (var i = 0; i < snapshot1.docs.length; i++) {
-    if (snapshot1.docs[i].id != snapshot2.docs[i].id) {
-      return false;
-    }
-
-    if (!_unorderedDeepEquality.equals(
-        snapshot1.docs[i].data(), snapshot2.docs[i].data())) {
-      return false;
-    }
-  }
-  return true;
+  FakeQueryWithParent? get parentQuery => _parentQuery;
 }
